requirements.txt
feedparser
python-dotenv
schedule
beautifulsoup4
googletrans==4.0.0-rc1
